requirements.txt
pyyaml
pillow
watchdog  # For auto-reload script